requirements.txt
python-pptx
instaloader
